refactor(core): build working
First working version to build qTox. Had to disable some tests though.
