+++ src/widget/widget.cpp
@@ -52,6 +52,7 @@
 #include "core/core.h"
 #include "core/coreav.h"
 #include "core/corefile.h"
+#include "core/groupinvite.h"
 #include "src/friendlist.h"
 #include "src/grouplist.h"
 #include "src/model/chathistory.h"
@@ -59,7 +60,6 @@
 #include "src/model/chatroom/groupchatroom.h"
 #include "src/model/friend.h"
 #include "src/model/group.h"
-#include "src/model/groupinvite.h"
 #include "src/model/profile/profileinfo.h"
 #include "src/model/status.h"
 #include "src/net/updatecheck.h"
